requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==23.0.0
sqlalchemy==2.0.36
alembic==1.13.3
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator>=2.0.0
redis==5.1.0
boto3==1.35.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
aiosqlite==0.20.0
